Create func to resize all imgs on a directory & add unit tests
The idea behind resizing the pictures is to obtain more accurate
results when looking for people on a given image. When the quality
of an image is increased considerably, the accuracy of the algoritm
decreases.

More tests are needed to see if this could be considered as a final
solution for the accuracy problems.