Solve issue of having multiple non sense paths on a single person
facesMatches is a list which stores the names of the people that
was found on a given image. Since Python manages the memory as
reference objects, it is necesary to clean() the list at the
end of each iteration. This way the information from the n-th
iteration wouldn't be carried to the (n+1)-th iteration.